Add option to to restock all shops
