+++ src/features/island/buildings/components/building/market/Restock.tsx
@@ -1,4 +1,4 @@
-import React, { useContext, useState } from "react";
+import React, { useContext } from "react";
 import { Button } from "components/ui/Button";
 import { ITEM_DETAILS } from "features/game/types/images";
 import { Context } from "features/game/GameProvider";
@@ -7,19 +7,10 @@ import { ModalContext } from "features/game/components/modal/ModalProvider";
 import stockIcon from "assets/icons/stock.webp";
 import { gameAnalytics } from "lib/gameAnalytics";
 import { useAppTranslation } from "lib/i18n/useAppTranslations";
-import { NPC_WEARABLES } from "lib/npcs";
 import { BB_TO_GEM_RATIO } from "features/game/types/game";
-import { hasFeatureAccess } from "lib/flags";
 import { TimerDisplay } from "features/retreat/components/auctioneer/AuctionDetails";
-import { useCountdown } from "lib/utils/hooks/useCountdown";
-import {
-  canRestockShipment,
-  nextShipmentAt,
-  SHIPMENT_STOCK,
-} from "features/game/events/landExpansion/shipmentRestocked";
+
 import { Label } from "components/ui/Label";
-import { Modal } from "components/ui/Modal";
-import { Panel } from "components/ui/Panel";
 import confetti from "canvas-confetti";
 import { Box } from "components/ui/Box";
 import Decimal from "decimal.js-light";
@@ -28,98 +19,98 @@ import { TREASURE_TOOLS, WORKBENCH_TOOLS } from "features/game/types/tools";
 import { SEEDS } from "features/game/types/seeds";
 import { CROP_LIFECYCLE } from "features/island/plots/lib/plant";
 
-export const Restock: React.FC = () => {
-  const { t } = useAppTranslation();
-  const { gameService } = useContext(Context);
-  const [gameState] = useActor(gameService);
+// export const Restock: React.FC = () => {
+//   const { t } = useAppTranslation();
+//   const { gameService } = useContext(Context);
+//   const [gameState] = useActor(gameService);
 
-  const [showConfirm, setShowConfirm] = useState(false);
+//   const [showConfirm, setShowConfirm] = useState(false);
 
-  const hasGemExperiment = hasFeatureAccess(
-    gameState.context.state,
-    "GEM_BOOSTS",
-  );
+//   const hasGemExperiment = hasFeatureAccess(
+//     gameState.context.state,
+//     "GEM_BOOSTS",
+//   );
 
-  const shipmentAt = useCountdown(
-    nextShipmentAt({ game: gameState.context.state }),
-  );
+//   const shipmentAt = useCountdown(
+//     nextShipmentAt({ game: gameState.context.state }),
+//   );
 
-  const { ...shipmentTime } = shipmentAt;
-  const shipmentIsReady = canRestockShipment({ game: gameState.context.state });
+//   const { ...shipmentTime } = shipmentAt;
+//   const shipmentIsReady = canRestockShipment({ game: gameState.context.state });
 
-  const showShipment = hasGemExperiment && shipmentIsReady;
+//   const showShipment = hasGemExperiment && shipmentIsReady;
 
-  if (showShipment) {
-    return (
-      <>
-        <Modal show={showConfirm} onHide={() => setShowConfirm(false)}>
-          <Panel className="sm:w-4/5 m-auto">
-            <ExperimentRestockModal onClose={() => setShowConfirm(false)} />
-          </Panel>
-        </Modal>
-        <Button className="relative" onClick={() => setShowConfirm(true)}>
-          <div className="flex items-center h-4 ">
-            <p>{t("restock")}</p>
-            <img src={stockIcon} className="h-6 absolute right-1 top-0" />
-          </div>
-        </Button>
-      </>
-    );
-  }
+//   if (showShipment) {
+//     return (
+//       <>
+//         <Modal show={showConfirm} onHide={() => setShowConfirm(false)}>
+//           <Panel className="sm:w-4/5 m-auto">
+//             <ExperimentRestockModal onClose={() => setShowConfirm(false)} />
+//           </Panel>
+//         </Modal>
+//         <Button className="relative" onClick={() => setShowConfirm(true)}>
+//           <div className="flex items-center h-4 ">
+//             <p>{t("restock")}</p>
+//             <img src={stockIcon} className="h-6 absolute right-1 top-0" />
+//           </div>
+//         </Button>
+//       </>
+//     );
+//   }
 
-  if (hasGemExperiment) {
-    return (
-      <>
-        <div className="flex justify-center items-center">
-          {/* <img src={stockIcon} className="h-5 mr-1" /> */}
-          <p className="text-xxs">{t("gems.nextFreeShipment")}</p>
-        </div>
-        <div className="flex justify-center items-center">
-          <img src={stockIcon} className="h-5 mr-1" />
-          <TimerDisplay time={shipmentTime} />
-        </div>
-        <div className="my-1 flex flex-col mb-1 flex-1 items-center justify-end">
-          <div className="flex items-center"></div>
-        </div>
-        <Button className="mt-1 relative" onClick={() => setShowConfirm(true)}>
-          <div className="flex items-center h-4 ">
-            <p>{t("restock")}</p>
-            <img
-              src={ITEM_DETAILS["Gem"].image}
-              className="h-5 absolute right-1 top-1"
-            />
-          </div>
-        </Button>
+//   if (hasGemExperiment) {
+//     return (
+//       <>
+//         <div className="flex justify-center items-center">
+//           {/* <img src={stockIcon} className="h-5 mr-1" /> */}
+//           <p className="text-xxs">{t("gems.nextFreeShipment")}</p>
+//         </div>
+//         <div className="flex justify-center items-center">
+//           <img src={stockIcon} className="h-5 mr-1" />
+//           <TimerDisplay time={shipmentTime} />
+//         </div>
+//         <div className="my-1 flex flex-col mb-1 flex-1 items-center justify-end">
+//           <div className="flex items-center"></div>
+//         </div>
+//         <Button className="mt-1 relative" onClick={() => setShowConfirm(true)}>
+//           <div className="flex items-center h-4 ">
+//             <p>{t("restock")}</p>
+//             <img
+//               src={ITEM_DETAILS["Gem"].image}
+//               className="h-5 absolute right-1 top-1"
+//             />
+//           </div>
+//         </Button>
 
-        <Modal show={showConfirm} onHide={() => setShowConfirm(false)}>
-          <Panel className="sm:w-4/5 m-auto" bumpkinParts={NPC_WEARABLES.betty}>
-            <ExperimentRestockModal onClose={() => setShowConfirm(false)} />
-          </Panel>
-        </Modal>
-      </>
-    );
-  }
+//         <Modal show={showConfirm} onHide={() => setShowConfirm(false)}>
+//           <Panel className="sm:w-4/5 m-auto" bumpkinParts={NPC_WEARABLES.betty}>
+//             <ExperimentRestockModal onClose={() => setShowConfirm(false)} />
+//           </Panel>
+//         </Modal>
+//       </>
+//     );
+//   }
 
-  return (
-    <>
-      <Button className="mt-1 relative" onClick={() => setShowConfirm(true)}>
-        <div className="flex items-center h-4 ">
-          <p>{t("restock")}</p>
-          <img
-            src={ITEM_DETAILS["Gem"].image}
-            className="h-5 absolute right-1 top-1"
-          />
-        </div>
-      </Button>
+//   return (
+//     <>
+//       <Button className="mt-1 relative" onClick={() => setShowConfirm(true)}>
+//         <div className="flex items-center h-4 ">
+//           <p>{t("restock")}</p>
+//           <img
+//             src={ITEM_DETAILS["Gem"].image}
+//             className="h-5 absolute right-1 top-1"
+//           />
+//         </div>
+//       </Button>
 
-      <Modal show={showConfirm} onHide={() => setShowConfirm(false)}>
-        <Panel className="sm:w-4/5 m-auto" bumpkinParts={NPC_WEARABLES.betty}>
-          <RestockModal onClose={() => setShowConfirm(false)} />
-        </Panel>
-      </Modal>
-    </>
-  );
-};
+//       <Modal show={showConfirm} onHide={() => setShowConfirm(false)}>
+//         <Panel className="sm:w-4/5 m-auto" bumpkinParts={NPC_WEARABLES.betty}>
+//           <RestockModal onClose={() => setShowConfirm(false)} />
+//         </Panel>
+//       </Modal>
+//     </>
+//   );
+// };
 
 interface RestockModalProps {
   onClose: () => void;
@@ -130,7 +121,7 @@ interface RestockModalProps {
   };
 }
 
-const RestockModal: React.FC<RestockModalProps> = ({
+export const RestockModal: React.FC<RestockModalProps> = ({
   onClose,
   shipmentTime,
 }) => {
@@ -267,154 +258,154 @@ const RestockModal: React.FC<RestockModalProps> = ({
   );
 };
 
-const ExperimentRestockModal: React.FC<{ onClose: () => void }> = ({
-  onClose,
-}) => {
-  const { t } = useAppTranslation();
+// const ExperimentRestockModal: React.FC<{ onClose: () => void }> = ({
+//   onClose,
+// }) => {
+//   const { t } = useAppTranslation();
 
-  const { gameService, showAnimations } = useContext(Context);
-  const [gameState] = useActor(gameService);
+//   const { gameService, showAnimations } = useContext(Context);
+//   const [gameState] = useActor(gameService);
 
-  const hasGemExperiment = hasFeatureAccess(
-    gameState.context.state,
-    "GEM_BOOSTS",
-  );
-  const shipmentIsReady = canRestockShipment({ game: gameState.context.state });
-  const showShipment = hasGemExperiment && shipmentIsReady;
+//   const hasGemExperiment = hasFeatureAccess(
+//     gameState.context.state,
+//     "GEM_BOOSTS",
+//   );
+//   const shipmentIsReady = canRestockShipment({ game: gameState.context.state });
+//   const showShipment = hasGemExperiment && shipmentIsReady;
 
-  const replenish = () => {
-    gameService.send("shipment.restocked");
+//   const replenish = () => {
+//     gameService.send("shipment.restocked");
 
-    if (showAnimations) confetti();
-    onClose();
-  };
+//     if (showAnimations) confetti();
+//     onClose();
+//   };
 
-  const shipmentAt = useCountdown(
-    nextShipmentAt({ game: gameState.context.state }),
-  );
+//   const shipmentAt = useCountdown(
+//     nextShipmentAt({ game: gameState.context.state }),
+//   );
 
-  const getShipmentAmount = (item: StockableName, amount: number): Decimal => {
-    const totalStock = INITIAL_STOCK(gameState.context.state)[item];
-    const remainingStock =
-      gameState.context.state.stock[item] ?? new Decimal(0);
-    // If shipment amount will exceed total stock
-    if (remainingStock.add(amount).gt(totalStock)) {
-      // return the difference between total and remaining stock
-      return totalStock.sub(remainingStock);
-    } else {
-      // else return shipment stock
-      return new Decimal(amount);
-    }
-  };
+//   const getShipmentAmount = (item: StockableName, amount: number): Decimal => {
+//     const totalStock = INITIAL_STOCK(gameState.context.state)[item];
+//     const remainingStock =
+//       gameState.context.state.stock[item] ?? new Decimal(0);
+//     // If shipment amount will exceed total stock
+//     if (remainingStock.add(amount).gt(totalStock)) {
+//       // return the difference between total and remaining stock
+//       return totalStock.sub(remainingStock);
+//     } else {
+//       // else return shipment stock
+//       return new Decimal(amount);
+//     }
+//   };
 
-  const restockTools = Object.entries(SHIPMENT_STOCK)
-    .filter((item) => item[0] in { ...WORKBENCH_TOOLS, ...TREASURE_TOOLS })
-    .filter(([item, amount]) => {
-      const shipmentAmount = getShipmentAmount(item as StockableName, amount);
-      return shipmentAmount.gt(0);
-    });
+//   const restockTools = Object.entries(SHIPMENT_STOCK)
+//     .filter((item) => item[0] in { ...WORKBENCH_TOOLS, ...TREASURE_TOOLS })
+//     .filter(([item, amount]) => {
+//       const shipmentAmount = getShipmentAmount(item as StockableName, amount);
+//       return shipmentAmount.gt(0);
+//     });
 
-  const restockSeeds = Object.entries(SHIPMENT_STOCK)
-    .filter((item) => item[0] in SEEDS())
-    .filter(([item, amount]) => {
-      const shipmentAmount = getShipmentAmount(item as StockableName, amount);
-      return shipmentAmount.gt(0);
-    });
+//   const restockSeeds = Object.entries(SHIPMENT_STOCK)
+//     .filter((item) => item[0] in SEEDS())
+//     .filter(([item, amount]) => {
+//       const shipmentAmount = getShipmentAmount(item as StockableName, amount);
+//       return shipmentAmount.gt(0);
+//     });
 
-  const restockIsEmpty = [...restockSeeds, ...restockTools].length <= 0;
+//   const restockIsEmpty = [...restockSeeds, ...restockTools].length <= 0;
 
-  if (showShipment) {
-    return (
-      <>
-        {restockIsEmpty ? (
-          <div className="p-1">
-            <Label type="danger" className="mb-2" icon={stockIcon}>
-              {t("gems.noShipment")}
-            </Label>
-            <p className="text-sm mb-2">{t("gems.buyStock")}</p>
-          </div>
-        ) : (
-          <div className="p-1">
-            <Label type="default" className="mb-2" icon={stockIcon}>
-              {t("gems.shipment.arrived")}
-            </Label>
-            <p className="text-sm mb-2">{t("gems.shipment.success")}</p>
-          </div>
-        )}
-        <div className="mt-1 h-40 overflow-y-auto overflow-x-hidden scrollable pl-1">
-          {restockTools.length > 0 && (
-            <Label
-              icon={ITEM_DETAILS.Axe.image}
-              type="default"
-              className="ml-2 mb-1"
-            >
-              {t("tools")}
-            </Label>
-          )}
-          <div className="flex flex-wrap mb-2">
-            {restockTools.map(([item, amount]) => {
-              const shipmentAmount = getShipmentAmount(
-                item as StockableName,
-                amount,
-              );
-              return (
-                <Box
-                  key={item}
-                  count={shipmentAmount}
-                  image={ITEM_DETAILS[item as StockableName].image}
-                />
-              );
-            })}
-          </div>
-          {restockSeeds.length > 0 && (
-            <Label
-              icon={CROP_LIFECYCLE.Sunflower.seed}
-              type="default"
-              className="ml-2 mb-1"
-            >
-              {t("seeds")}
-            </Label>
-          )}
-          <div className="flex flex-wrap mb-2">
-            {restockSeeds.map(([item, amount]) => {
-              const shipmentAmount = getShipmentAmount(
-                item as StockableName,
-                amount,
-              );
-              return (
-                <Box
-                  key={item}
-                  count={shipmentAmount}
-                  image={ITEM_DETAILS[item as StockableName].image}
-                />
-              );
-            })}
-          </div>
-        </div>
-        {!restockIsEmpty && (
-          <>
-            <p className="text-xs p-1 pb-1.5 italic">
-              {t("gems.restockToMaxStock")}
-            </p>
-            <p className="text-xs p-1 pb-1.5 italic">
-              {`(${t("gems.shipment.useGems")})`}
-            </p>
-          </>
-        )}
+//   if (showShipment) {
+//     return (
+//       <>
+//         {restockIsEmpty ? (
+//           <div className="p-1">
+//             <Label type="danger" className="mb-2" icon={stockIcon}>
+//               {t("gems.noShipment")}
+//             </Label>
+//             <p className="text-sm mb-2">{t("gems.buyStock")}</p>
+//           </div>
+//         ) : (
+//           <div className="p-1">
+//             <Label type="default" className="mb-2" icon={stockIcon}>
+//               {t("gems.shipment.arrived")}
+//             </Label>
+//             <p className="text-sm mb-2">{t("gems.shipment.success")}</p>
+//           </div>
+//         )}
+//         <div className="mt-1 h-40 overflow-y-auto overflow-x-hidden scrollable pl-1">
+//           {restockTools.length > 0 && (
+//             <Label
+//               icon={ITEM_DETAILS.Axe.image}
+//               type="default"
+//               className="ml-2 mb-1"
+//             >
+//               {t("tools")}
+//             </Label>
+//           )}
+//           <div className="flex flex-wrap mb-2">
+//             {restockTools.map(([item, amount]) => {
+//               const shipmentAmount = getShipmentAmount(
+//                 item as StockableName,
+//                 amount,
+//               );
+//               return (
+//                 <Box
+//                   key={item}
+//                   count={shipmentAmount}
+//                   image={ITEM_DETAILS[item as StockableName].image}
+//                 />
+//               );
+//             })}
+//           </div>
+//           {restockSeeds.length > 0 && (
+//             <Label
+//               icon={CROP_LIFECYCLE.Sunflower.seed}
+//               type="default"
+//               className="ml-2 mb-1"
+//             >
+//               {t("seeds")}
+//             </Label>
+//           )}
+//           <div className="flex flex-wrap mb-2">
+//             {restockSeeds.map(([item, amount]) => {
+//               const shipmentAmount = getShipmentAmount(
+//                 item as StockableName,
+//                 amount,
+//               );
+//               return (
+//                 <Box
+//                   key={item}
+//                   count={shipmentAmount}
+//                   image={ITEM_DETAILS[item as StockableName].image}
+//                 />
+//               );
+//             })}
+//           </div>
+//         </div>
+//         {!restockIsEmpty && (
+//           <>
+//             <p className="text-xs p-1 pb-1.5 italic">
+//               {t("gems.restockToMaxStock")}
+//             </p>
+//             <p className="text-xs p-1 pb-1.5 italic">
+//               {`(${t("gems.shipment.useGems")})`}
+//             </p>
+//           </>
+//         )}
 
-        <div className="flex">
-          <Button className="mr-1" onClick={onClose}>
-            {t("close")}
-          </Button>
-          <Button onClick={replenish} disabled={restockIsEmpty}>
-            {t("restock")}
-          </Button>
-        </div>
-      </>
-    );
-  }
+//         <div className="flex">
+//           <Button className="mr-1" onClick={onClose}>
+//             {t("close")}
+//           </Button>
+//           <Button onClick={replenish} disabled={restockIsEmpty}>
+//             {t("restock")}
+//           </Button>
+//         </div>
+//       </>
+//     );
+//   }
 
-  const { ...shipmentTime } = shipmentAt;
+//   const { ...shipmentTime } = shipmentAt;
 
-  return <RestockModal shipmentTime={shipmentTime} onClose={onClose} />;
-};
+//   return <RestockModal shipmentTime={shipmentTime} onClose={onClose} />;
+// };
