+++ src/features/island/buildings/components/building/workBench/components/Tools.tsx
@@ -13,7 +13,6 @@ import {
   LOVE_ANIMAL_TOOLS,
 } from "features/game/types/tools";
 import { getKeys } from "features/game/types/craftables";
-import { Restock } from "features/island/buildings/components/building/market/Restock";
 import { SplitScreenView } from "components/ui/SplitScreenView";
 import { CraftingRequirements } from "components/ui/layouts/CraftingRequirements";
 import { makeBulkBuyTools } from "../../market/lib/makeBulkBuyAmount";
@@ -27,7 +26,7 @@ import { capitalize } from "lib/utils/capitalize";
 import { IslandType, LoveAnimalItem } from "features/game/types/game";
 import { getToolPrice } from "features/game/events/landExpansion/craftTool";
 import { hasFeatureAccess } from "lib/flags";
-import { EnhancedRestock } from "../../market/EnhancedRestock";
+import { Restock } from "../../market/RestockButton";
 
 const isLoveAnimalTool = (
   toolName: WorkbenchToolName | LoveAnimalItem,
@@ -133,9 +132,7 @@ export const Tools: React.FC = () => {
     }
 
     if (stock.equals(0)) {
-      if (hasFeatureAccess(state, "ENHANCED_RESTOCK"))
-        return <EnhancedRestock npc={"blacksmith"} />;
-      return <Restock />;
+      return <Restock npc={"blacksmith"} />;
     }
 
     return (
