+++ src/features/world/ui/beach/treasure_shop/TreasureShopBuy.tsx
@@ -10,7 +10,6 @@ import { getKeys } from "features/game/types/decorations";
 import { ITEM_DETAILS } from "features/game/types/images";
 import { TREASURE_TOOLS, TreasureToolName } from "features/game/types/tools";
 import { makeBulkBuyTools } from "features/island/buildings/components/building/market/lib/makeBulkBuyAmount";
-import { Restock } from "features/island/buildings/components/building/market/Restock";
 import { useAppTranslation } from "lib/i18n/useAppTranslations";
 import {
   TREASURE_COLLECTIBLE_ITEM,
@@ -34,8 +33,7 @@ import lightning from "assets/icons/lightning.png";
 import { getToolPrice } from "features/game/events/landExpansion/craftTool";
 import { Keys } from "features/game/types/game";
 import { isMobile } from "mobile-device-detect";
-import { hasFeatureAccess } from "lib/flags";
-import { EnhancedRestock } from "features/island/buildings/components/building/market/EnhancedRestock";
+import { Restock } from "features/island/buildings/components/building/market/RestockButton";
 
 interface ToolContentProps {
   selectedName: TreasureToolName;
@@ -90,11 +88,7 @@ const ToolContent: React.FC<ToolContentProps> = ({ selectedName }) => {
       actionView={
         <>
           {stock.equals(0) ? (
-            hasFeatureAccess(state, "ENHANCED_RESTOCK") ? (
-              <EnhancedRestock npc={"jafar"} />
-            ) : (
-              <Restock />
-            )
+            <Restock npc={"jafar"} />
           ) : (
             <div className="flex space-x-1 sm:space-x-0 sm:space-y-1 sm:flex-col w-full">
               <Button
